Cleanup for better testing; Improved TF so that it can fulfill Emitter requirements

+++ src/common/particle_system/cpu_i_particle_system.cpp
@@ -46,7 +46,7 @@ bool CpuIParticleSystem::AddModule(CpuIModule* psModule)
 	return true;
 }
 
-size_t CpuIParticleSystem::GetCurrentParticles() const
+uint32_t CpuIParticleSystem::GetCurrentParticles() const
 {
 	return mNumParticles;
 }
